% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{add_index_densities}
\alias{add_index_densities}
\title{Extract the numbers from the growing density matrix for this specific sample.}
\usage{
add_index_densities(composite, first, second, sname)
}
\arguments{
\item{composite}{Table to build up.}

\item{first}{First piece to add to the composite.}

\item{second}{Second piece to add to the composite.}

\item{sname}{Sample name.}
}
\description{
Then add them and make a new matrix with these numbers.
}
