% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{barplot_matrices}
\alias{barplot_matrices}
\title{Make a bar plot for every type of data returned by create_matrices.}
\usage{
barplot_matrices(retlist, bar_title = "")
}
\arguments{
\item{retlist}{Result from create_matrices() with all the fun stuff to plot.}

\item{bar_title}{Title for the plot.}
}
\description{
The function create_matrices() creates two lists containing the various
categorizations of the data.  This takes that information and makes a bar
plot for every element in those lists.
}
