% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.r
\name{count_mutation_direction}
\alias{count_mutation_direction}
\title{Count up how many reads are on the forward and reverse strands.}
\usage{
count_mutation_direction(identical, changed, sequencer, verbose = FALSE)
}
\arguments{
\item{identical}{Set of reads identical to the template.}

\item{changed}{Reads deemed to be from the RT.}

\item{sequencer}{Reads deemed to be from the sequencer.}

\item{verbose}{Print some information while running.}
}
\value{
List of how many forward and reverse reads are in the data.
}
\description{
Count up how many reads are on the forward and reverse strands.
}
