% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.r
\name{count_mutation_types}
\alias{count_mutation_types}
\title{Count up all sorts of fun mutation types in the data!}
\usage{
count_mutation_types(
  strict_rt,
  strict_sequencer,
  min_indexes = 3,
  verbose = FALSE
)
}
\arguments{
\item{strict_rt}{Indexes from RT errors.}

\item{strict_sequencer}{Indexes from sequencer errors.}

\item{min_indexes}{The minimum number of indexes required to deem a given
error as 'real'.}

\item{verbose}{Print some information about the run.}
}
\value{
Tables of the various mutation types.
}
\description{
This is my favorite function in this!  It counts up all sorts of fun mutation
types in the data and gives back tables of the results.
}
