% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{create_matrices}
\alias{create_matrices}
\title{Given a samples sheet with some metadata, create a big pile of matrices
describing the data.}
\usage{
create_matrices(
  sample_sheet = "sample_sheets/all_samples.xlsx",
  ident_column = "identtable",
  mut_column = "mutationtable",
  min_reads = NULL,
  min_indexes = NULL,
  min_sequencer = 10,
  min_position = NULL,
  max_position = NULL,
  max_mutations_per_read = NULL,
  prune_n = TRUE,
  excel = NULL,
  verbose = TRUE,
  plot_order = c("dna_control", "dna_low", "dna_high", "rna_control", "rna_low",
    "rna_high"),
  savefile = "sample_tables.rda",
  overwrite = FALSE
)
}
\arguments{
\item{sample_sheet}{xlsx/csv/whatever file of metadata.}

\item{ident_column}{Column containing the files of reads identical to the
template.}

\item{mut_column}{Column containing the files of reads not identical to the
template.}

\item{min_reads}{Filter for the minimum number of reads / index.}

\item{min_indexes}{Filter for the minimum numer of indexes / mutation.}

\item{min_sequencer}{Filter defining the minimum number of reads when looking
for sequencer-based mutations.}

\item{min_position}{Filter against weird data at the beginning of the
template.}

\item{max_position}{Filter against weird data at the end.}

\item{max_mutations_per_read}{Drop reads with more than this number of mutations.}

\item{prune_n}{Remove mutations which are something to N.}

\item{excel}{Save data to this file.}

\item{verbose}{Print information about the running of this function?}

\item{plot_order}{When plotting, force the order to this.}

\item{savefile}{Save the timeconsuming portion to this file.}

\item{overwrite}{If the savefile exists, overwrite it?}
}
\value{
List with lots of matrices of the resulting data.
}
\description{
Given a samples sheet with some metadata, create a big pile of matrices
describing the data.
}
