% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_errRt.r
\docType{package}
\name{errRt}
\alias{errRt}
\title{errRt: Some functions for analyzing sequencing data intended to quantify RT error rates.}
\description{
errRt: Some functions for analyzing sequencing data intended to quantify RT error rates.
}
