% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_mutations.r
\name{expand_mutation_string}
\alias{expand_mutation_string}
\title{Combine the columns describing a mutation into a single column and categorize.}
\usage{
expand_mutation_string(t)
}
\arguments{
\item{t}{Table from readr.}
}
\value{
The same table with some new columns describing the mutations therein.
}
\description{
Given a table with columns including the position, mutation type, reference
nt at that position, and product nt at that position; create a single column
from it, standardize it, and make some categories describing each mutation.
These columns currently include:  'mt': X_Y telling that this is a mutation
from X to Y, 'transition_transversion': this is a transition or transversion
mismatch, 'strong_weak': this mismatch went from strong->weak, weak->strong,
weak->weak, or strong->strong.  undef is used in the case of indels.
}
