% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_max_mutations}
\alias{filter_max_mutations}
\title{Filter away reads which have > n mutations in them.}
\usage{
filter_max_mutations(
  chng,
  mutation_df,
  max_mutations_per_read = 10,
  verbose = FALSE
)
}
\arguments{
\item{chng}{Data set of reads which are not identical to the template.}

\item{mutation_df}{Data frame of readids and their number of mutations.}

\item{max_mutations_per_read}{What it says on the tin.}

\item{verbose}{Print some information while running?}
}
\value{
The remaining reads and a little summary information.
}
\description{
A subset of the reads have a weirdly large number of differences to the
template.  These add confusing noise to the data, however removing them may
take away meaningful noise from the data.  So this function may prove to be
awesome or terrible, depending on one's perspective.
}
