% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_max_position}
\alias{filter_max_position}
\title{Filter away reads after position x.}
\usage{
filter_max_position(chng, max_position = 176, verbose = FALSE)
}
\arguments{
\item{chng}{Data set of reads which are not identical to the template.}

\item{max_position}{Position after which to filter.}

\item{verbose}{Print some information while running?}
}
\value{
The remaining reads and a little summary information.
}
\description{
The error rate in the data is crazy-high in the immediate vicinity of the
RT/PCR primers.  Thus we exclude reads with mutations before this position.
}
