% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_min_position}
\alias{filter_min_position}
\title{Filter away reads before position x.}
\usage{
filter_min_position(chng, min_position = 24, verbose = FALSE)
}
\arguments{
\item{chng}{Data set of reads which are not identical to the template.}

\item{min_position}{Position before which to filter.}

\item{verbose}{Print some information while running?}
}
\value{
The remaining reads and a little summary information.
}
\description{
The error rate in the data is crazy-high in the immediate vicinity of the
RT/PCR primers.  Thus we exclude reads with mutations before this position.
}
