% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_ns}
\alias{filter_ns}
\title{Filter away reads which contain Ns.}
\usage{
filter_ns(chng, verbose = FALSE)
}
\arguments{
\item{chng}{Data set of reads which are not identical to the template.}

\item{verbose}{Print some information while running?}
}
\value{
The remaining reads and a little summary information.
}
\description{
It is unlikely that we will make much sense out of reads with N.  So toss
em!
}
