% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_proportion}
\alias{filter_proportion}
\title{Filter out thing which are of an insufficient proportion.}
\usage{
filter_proportion(chng, ident, min_proportion = 0.5, verbose = FALSE)
}
\arguments{
\item{chng}{Table of changed reads.}

\item{ident}{Table of identical reads.}

\item{min_proportion}{Minimum proportion of changed/identical reads/index.}

\item{verbose}{Print while running.}
}
\description{
This is not currently used, but was something I wanted to consider.
}
