% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_pruned_indexes}
\alias{filter_pruned_indexes}
\title{Filter away reads which are associated with indexes that have been deemed
insufficient.  These 'bad' indexes were defined by prune_indexes().}
\usage{
filter_pruned_indexes(
  chng,
  ident,
  wanted_indexes,
  min_reads = 3,
  verbose = FALSE
)
}
\arguments{
\item{chng}{Data set of reads which are not identical to the template.}

\item{ident}{Data set of reads which are identical to the template.}

\item{wanted_indexes}{Character vector of indexes which have sufficient
evidence.}

\item{min_reads}{Used when verbose to remind us of the number of
reads/index.}

\item{verbose}{Print some information while running?}
}
\value{
The remaining reads and a little summary information.
}
\description{
At least in theory we should have at least x reads per index before that
index is deemed interesting.  prune_indexes() creates a list of indexes which
have more than this x reads per index and provides it to this.  This then
removes the rest.  Caveat: this filter works on both the mutated and
identical data.
}
