% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{make_index_table}
\alias{make_index_table}
\title{Given the large table of index counts for one sample, bring them together into a small matrix.}
\usage{
make_index_table(composite, piece, sname)
}
\arguments{
\item{composite}{The matrix to build.}

\item{piece}{The piece to add.}

\item{sname}{Sample name.}
}
\description{
Given the large table of index counts for one sample, bring them together into a small matrix.
}
