% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{matrices_from_tables}
\alias{matrices_from_tables}
\title{Given the many tables provided by quantify, merge them into matrices of error
types.}
\usage{
matrices_from_tables(tables, samples, verbose = FALSE)
}
\arguments{
\item{tables}{The big pile of error tables.}

\item{samples}{The set of samples in the data.}

\item{verbose}{Print some information while running.}
}
\value{
Matrices with error types as rows, samples as columns.
}
\description{
Given the many tables provided by quantify, merge them into matrices of error
types.
}
