% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.r
\name{normalize_matrices}
\alias{normalize_matrices}
\title{Invoke some normalization methods on the various matrices from
matrices_from_tables().}
\usage{
normalize_matrices(matrices, reads_per_sample, indexes_per_sample)
}
\arguments{
\item{matrices}{Tables provided by matrices_from_tables().}

\item{reads_per_sample}{Numeric vector defining how many reads survived
filtering for each sample.}

\item{indexes_per_sample}{Numeric vector defining how many indexes survived
filtering for each sample.}
}
\value{
Normalized matrices
}
\description{
Invoke some normalization methods on the various matrices from
matrices_from_tables().
}
