% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_index_density}
\alias{plot_index_density}
\title{Make a density plot of how many indexes are associated with each sample.}
\usage{
plot_index_density(df, max = 20)
}
\arguments{
\item{df}{Table of indexes/sample.}

\item{max}{Maximum number of indexes to plot.}
}
\value{
Density plot!
}
\description{
Make a density plot of how many indexes are associated with each sample.
}
