% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{prune_indexes}
\alias{prune_indexes}
\title{Summarize the mutant/identical data with respect to the number of
reads/index.}
\usage{
prune_indexes(chng, ident, min_reads = 3, verbose = TRUE)
}
\arguments{
\item{chng}{The result of read_tsv() on a file containing a mutation table.}

\item{ident}{The result of read_tsv() on a file containing identical reads.}

\item{min_reads}{Minimum read / index filter.}

\item{verbose}{Print information about this while it runs?}
}
\value{
List with the summary of the numbers of reads observed and the
 indexes kept.  The list of indexes kept may just be 'all'.
}
\description{
This should provide a table of how many reads/index are (a)identical, (b)contain
mutations, and the sum of (a + b).  If a minimum number of reads is requested
(e.g. min_reads is a number), return the list of indexes which have at least
that many reads.  This set of indexes may be used in other contexts to limit
the data.
}
