% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.r
\name{quantify_parsed}
\alias{quantify_parsed}
\title{Quantify the tables of reads identical/different to/from the template sequence.}
\usage{
quantify_parsed(
  changed = NULL,
  identical = NULL,
  min_reads = NULL,
  min_indexes = NULL,
  min_sequencer = 10,
  prune_n = TRUE,
  min_position = 24,
  max_position = 176,
  max_mutations_per_read = NULL,
  reencode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{changed}{File containing the set of changed reads/indexes, by default
named 'step4.txt.xz' by errrt.pm.}

\item{identical}{File containing the set of identical reads/indexes, by
default named 'step2_identical_reads.txt.xz' by errrt.pm.}

\item{min_reads}{Minimum number of reads for each index required to include
each index in the final result.}

\item{min_indexes}{Minimum number of indexes required to include a given
mutation in the final result.}

\item{min_sequencer}{Minimum number of total reads to consider an index as
associated with a sequencer-based error.}

\item{prune_n}{Remove mutations of a base to 'N'?}

\item{min_position}{Filter mutations before this position?}

\item{max_position}{Filter mutations after this position?}

\item{max_mutations_per_read}{If a read has more than this number of mutations, drop it.}

\item{reencode}{Rewrite the indexes on a different scale to save memory (6 bit instead of 2)}

\item{verbose}{Print information describing what is happening while this runs.}
}
\value{
List containing a bunch of summary information about the data.
}
\description{
The heavy lifting of locating reads which are/not identical to the template was
performed by errrt.pm and returns a series of compressed tables of identical read
ids and non-identical insertions, deletions, and mismatches.  This function is
intended to read those two files, gather the resulting data, and make
some sense of it.
}
