% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate.r
\name{sequencer_error}
\alias{sequencer_error}
\title{Calculate a lower limit error rate for the sequencer.}
\usage{
sequencer_error(data_summary)
}
\arguments{
\item{data_summary}{Result from create_matrices().}
}
\value{
Estimate of errors/nucleotide deemed to originate from the sequencer.
}
\description{
I say 'lower limit' because I assume my set of sequencer-based errors did not
detect the full set of error actually from the sequencer.  I chose to take
only the set with >= 10 reads / index for which only n (1) read(s) is a
mutant.  My hope is that this will avoid false positives, but it will also
limit the perceived sequencer error rate.
}
