% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.r
\name{write_legend_sheet}
\alias{write_legend_sheet}
\title{Write an initial sheet describing what to expect in the rest of the workbook.}
\usage{
write_legend_sheet(
  wb,
  sheet = "legend",
  min_reads = NULL,
  min_indexes = NULL,
  min_sequencer = 10,
  min_position = NULL,
  max_position = NULL,
  max_mutations_per_read = NULL,
  prune_n = NULL
)
}
\arguments{
\item{wb}{Workbook to write}

\item{sheet}{Worksheet to write}

\item{min_reads}{Passed down from create_matrices, printed in a table.}

\item{min_indexes}{Passed down from create_matrices, printed in a table.}

\item{min_sequencer}{Passed down from create_matrices, printed in a table.}

\item{min_position}{Passed down from create_matrices, printed in a table.}

\item{max_position}{Passed down from create_matrices, printed in a table.}

\item{max_mutations_per_read}{Passed down from create_matrices, printed in a table.}

\item{prune_n}{Passed down from create_matrices, printed in a table.}
}
\description{
Write an initial sheet describing what to expect in the rest of the workbook.
}
