% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.r
\name{write_matrices}
\alias{write_matrices}
\title{Make a pretty xlsx file of the various summary statistics and data from an RT
error rate experiment.}
\usage{
write_matrices(retlist, excel)
}
\arguments{
\item{retlist}{Result from create_matrices()}

\item{excel}{Excel filename.}
}
\description{
This takes the results of create_matrices() and hopefully prints some pretty
data about them.
}
