% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.r
\name{write_matrix_sheet}
\alias{write_matrix_sheet}
\title{Write out a worksheet containing one group of error rate matrices}
\usage{
write_matrix_sheet(
  retlst,
  wb,
  table_names,
  sheet = "raw",
  table_type = "matrices"
)
}
\arguments{
\item{retlst}{Set of results from create_matrices()}

\item{wb}{Workbook to write.}

\item{table_names}{The set of tables to write.}

\item{sheet}{Name of the sheet to write.}

\item{table_type}{I have a few different tables for each metric, raw, cpm,
cpmlength, etc.  Choose one.}
}
\description{
This function is responsible for writing out the actual data.  Most of the
logic in it is in an attempt to properly space the various tables and plots
apart.
}
