% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.r
\name{barplot_matrices}
\alias{barplot_matrices}
\title{Make a bar plot for every type of data returned by create_matrices.}
\usage{
barplot_matrices(summary)
}
\arguments{
\item{summary}{Result from create_matrices()}
}
\value{
List full of bar plots!
}
\description{
The function create_matrices() creates two lists containing the various
categorizations of the data.  This takes that information and makes a bar
plot for every element in those lists.
}
