% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_errRt.r
\docType{package}
\name{errRt}
\alias{errRt}
\title{errRt: Some functions for analyzing sequencing data intended to quantify RT error rates.}
\description{
This package (along with its perl counterpart) attempts to make it easier to
analyze high-throughput experiments which quantify error rates by using a
sample metadata worksheet to automate the filtering, classification, and
counting of errors in the sequencing data of a known template.
}
