% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_index_density}
\alias{plot_index_density}
\title{Make a density plot of how many indexes are associated with each sample.}
\usage{
plot_index_density(data, metadata, title = "")
}
\arguments{
\item{data}{Table of indexes/sample.}

\item{metadata}{Experiment metadata, used to acquire colors.}

\item{title}{Plot title.}
}
\value{
Density plot!
}
\description{
Make a density plot of how many indexes are associated with each sample.
}
